Trait resize_slice::SliceExt

pub trait SliceExt<T> {
    fn copy_from(&mut self, src: &[T]) -> usize where T: Copy;
    fn copy_inner(&mut self, src: usize, dst: usize, len: usize) where T: Copy;
}

Extension methods for slices

Required Methods

fn copy_from(&mut self, src: &[T]) -> usize where T: Copy

Copies the less of self.len() and src.len() from src into self, returning the amount of items copies.

fn copy_inner(&mut self, src: usize, dst: usize, len: usize) where T: Copy

Copies elements to another location within the slice, which may overlap.