Trait resize_slice::VecExt

pub trait VecExt<T> {
    unsafe fn uninitialized_resize(&mut self, new_len: usize);
    unsafe fn zeroed_resize(&mut self, new_len: usize);
}

Extension methods for vector types

Required Methods

unsafe fn uninitialized_resize(&mut self, new_len: usize)

Unsafely a vector to the specified size, without initializing the memory.

unsafe fn zeroed_resize(&mut self, new_len: usize)

Unsafely a vector to the specified size, zeroing the memory.